pyproject.toml
[tool.poetry]
name = "music-upgrader"
version = "0.1.0"
description = ""
authors = ["Joel Cochran <joel@cochrandigital.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mutagen = "^1.47.0"
python-dateutil = "^2.9.0.post0"
beets = "^2.0.0"
inflection = "^0.5.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
beets-yearfixer = "^0.0.5"
pylast = "^5.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
